pyproject.toml
[tool.poetry]
name = "rag-project"
version = "0.1.0"
description = "A production-ready RAG system with LangGraph, LangChain, MLflow, and ChromaDB"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.2.0"
langchain-community = "^0.2.0"
langchain-core = "^0.2.0"
langchain-text-splitters = "^0.2.0"
langchain-openai = "^0.1.0"
langchain-ollama = "^0.1.0"
langchain-chroma = "^0.1.0"
langgraph = "^0.2.0"
chromadb = "^0.4.22"
mlflow = "^2.10.0"
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
tiktoken = "^0.7.0"
pypdf = "^4.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.0"
black = "^24.0.0"
ruff = "^0.1.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html --cov-report=term"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
